% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagrams.R
\name{gbk_diagrams_cosmology}
\alias{gbk_diagrams_cosmology}
\title{Render all Cosmology diagrams}
\usage{
gbk_diagrams_cosmology(out = here::here("out/"))
}
\arguments{
\item{out}{}
}
\value{

}
\description{
Render all Cosmology diagrams
}
